Make a Kaplan-Meier plot by meta-dichotomizing a feature

A major plotting function that takes in a meta-analytic dataset and a feature-name and outputs a survival curve based on dataset-wise median dichotomization of the feature.

Usage

create.meta.prognosis.plot(expression.data, survival.data, feature.name, filename,
    xlab.label = "Time (Years)", ylab.label = "Fraction of Cohort", xlimits, xat,
    label.times, label.positions, label.cex = 1.2, main = NULL, main.cex = 3.0,
    rounding = 2, covariates = NULL, line.colours = c("blue", "red"), resolution = 2000,
    enable.warnings = FALSE, description = NULL);

Arguments

expression.data

A list of matrices, with each list element representing an expression matrix for a separate dataset

survival.data

A list of survival objects, with each list element representing a survobj for a separate dataset

feature.name

The name of the feature to be dichotomized

filename

The filename to be written

xlab.label

X Axis Title. Defaults to 'Time (Years)'

ylab.label

Y Axis Title. Defaults to 'Fraction of Cohort'

xlimits

The limits of the x (time) axis

xat

The locations of the x-axis labels

label.times

A vector of times at which to calculate the fraction of patients still at risk

label.positions

A vector of positions at which to print the fraction of patients still at risk

label.cex

Size of patients at risk text

main

The main title for the plot (space is reclaimed if NULL)

main.cex

Size of text for main plot title

rounding

How many decimal places of precision should be shown in printed survival statistics?

covariates

A data-frame (or matrix?) of variables to be controlled in the Cox model. If null, no adjustment is done. No interactions are fit.

line.colours

A vector of colours to be used in plotting

resolution

Figure resolution, defaults to 2000

enable.warnings

Print warnings if set to TRUE, defaults to FALSE

description

Short description of image/plot, defaults to NULL

Value

Returns a survival curve (written to file) and a four-element vector containing the survival results.

Author

Paul C. Boutros

See also

Examples

 # add examples