Utility function to make complete Kaplan-Meier plots

A utility function that makes a pretty Kaplan-Meier survival curve given survival data and patient grouping. This is a highly general function that does very little with the survival data itself.

Usage

make.survival.plot(survival.object, patient.groups, filename, xlimits, xat,
    label.positions, label.cex = 1.2, at.risk.low, at.risk.high,
    risk.labels = c("Low", "High"), main = NULL, main.cex = 3.0, rounding = 2,
    covariates = NULL, line.colours = c("blue", "red"), key = NULL,
    statistical.method = NA, x.statistic = 0.200, y.statistic = 0.610,
    resolution = 2000, height = 7.007874, width = 7.007874, size.units = 'in',
    enable.warnings = FALSE);

Arguments

survival.object

An object of class Surv -- patient ordering needs to be identical as for patient.groups

patient.groups

Grouping of patients (passed directly to coxph, so factors & continuous variables are okay)

filename

The filename to be written

xlimits

A two component vector listing the upper and lower bounds of the x-axis (time dimension)

xat

A vector listing the points at which x-axis labels should be drawn

label.positions

A vector of positions at which to print the fraction of patients still at risk

label.cex

Size of patients at risk text

at.risk.low

A vector containing the number of patients in the low-risk group at each label.position

at.risk.high

A vector containing the number of patients in the high-risk group at each label.position

risk.labels

A character(2) vector that gives the labels for the at-risk rows

main

The main title for the plot (space is reclaimed if NULL)

main.cex

Size of text for main plot title

rounding

How many decimal places of precision should be shown in printed survival statistics?

covariates

A data-frame (or matrix?) of variables to be controlled in the Cox model. If null, no adjustment is done. No interactions are fit.

line.colours

A vector of colours to be used in plotting

key

A list describing the legend one wants to add to the plot, defaults to no legend

statistical.method

A character indicating what statistical test should be done ('cox', 'logrank', 'ttest'). If NA, tries to choose intelligently

x.statistic

x position for statistical data, defaults to left

y.statistic

y position for statistical data, defaults to low

resolution

Figure resolution in dpi

height

Figure height in size.units

width

Figure width in size.units

size.units

Units of size for the figure

enable.warnings

Print warnings if set to TRUE, defaults to FALSE

Value

Creates a tiff file containing a full Kaplan-Meier curve, including number of patients at risk at multiple time-points.

Author

Paul C. Boutros

See also

Examples

# create some artificial data
survtime <- c(
  sample(1:75, 100, replace = TRUE),
  sample(1:99, 100, replace = TRUE)
  );

survstat <- c(
  sample( c(0,1,1), 100, replace = TRUE),
  sample( c(0,0,1), 100, replace = TRUE)
  )

survobj <- Surv(survtime, survstat);
groups <- as.factor(c( rep('A', 100), rep('B', 100) ));

# set some plotting parameters
xlimits <- c(0,100);
xat <- seq(0,100,25);
label.times <- seq(0,100,20);
label.positions <- 0.167 + 0.1605 * (1:length(label.times) - 1);

make.survival.plot(
  survival.object = survobj,
  patient.groups = groups,
  filename = 'survival_plot_simple.tiff',
  xlimits = xlimits,
  xat = xat,
  label.positions = label.positions,
  at.risk.low = calculate.number.at.risk(
    survobj = survobj[groups == 'B',],
    cut.points = label.times
    ),
  at.risk.high = calculate.number.at.risk(
    survobj = survobj[groups == 'A',],
    cut.points = label.times
    ),
  risk.labels = c('A = High', 'B = Low'),
  resolution = 100
  );
#> Warning: This function will be deprecated -- consider using create.survival.plot.R() instead
#> Warning: No TIFF support in this version of R
#> [1] 9.53513e-07 3.20000e-01 2.00000e-01 5.00000e-01